move key mixin to the mixin dir and update imports - also use it in secret-v2 model
