+++ ui/app/models/secret-v2.js
@@ -2,10 +2,11 @@ import DS from 'ember-data';
 import { computed } from '@ember/object';
 import { match } from '@ember/object/computed';
 import { expandAttributeMeta } from 'vault/utils/field-to-attrs';
+import KeyMixin from 'vault/mixins/key-mixin';
 
 const { attr, hasMany, belongsTo, Model } = DS;
 
-export default Model.extend({
+export default Model.extend(KeyMixin, {
   engine: belongsTo('secret-engine'),
   versions: hasMany('secret-v2-version', { async: false, inverse: null }),
   selectedVersion: belongsTo('secret-v2-version', { async: false, inverse: 'secret' }),
@@ -23,7 +24,6 @@ export default Model.extend({
     helpText:
       'Writes will only be allowed if the key’s current version matches the version specified in the cas parameter',
   }),
-  isFolder: match('id', /\/$/),
   fields: computed(function() {
     return expandAttributeMeta(this, ['maxVersions', 'casRequired']);
   }),
